+++ firebase-vscode/src/data-connect/terminal.ts
@@ -4,6 +4,7 @@ import vscode, { Disposable } from "vscode";
 import { checkLogin } from "../core/user";
 import { DATA_CONNECT_EVENT_NAME } from "../analytics";
 import { getSettings } from "../utils/settings";
+import { currentProjectId } from "../core/project";
 
 const environmentVariables: Record<string, string> = {};
 
@@ -26,12 +27,16 @@ export function runCommand(command: string) {
   // TODO: This fails if the interactive shell is not expecting a command, such
   // as when oh-my-zsh asking for (Y/n) to updates during startup.
   // Consider using an non-interactive shell.
+  if (currentProjectId.value) {
+    command = `${command} --project ${currentProjectId.value}`;
+  }
   terminal.sendText(command);
 }
 
 export function runTerminalTask(
   taskName: string,
   command: string,
+  presentationOptions: vscode.TaskPresentationOptions = { focus: true },
 ): Promise<string> {
   const type = "firebase-" + Date.now();
   return new Promise(async (resolve, reject) => {
@@ -43,20 +48,22 @@ export function runTerminalTask(
           resolve(`Successfully executed ${taskName} with command: ${command}`);
         } else {
           reject(
-            new Error(`Failed to execute ${taskName} with command: ${command}`),
+            new Error(
+              `{${e.exitCode}}: Failed to execute ${taskName} with command: ${command}`,
+            ),
           );
         }
       }
     });
-    const task = await vscode.tasks.executeTask(
-      new vscode.Task(
-        { type },
-        vscode.TaskScope.Workspace,
-        taskName,
-        "firebase",
-        new vscode.ShellExecution(command, executionOptions),
-      ),
+    const task = new vscode.Task(
+      { type },
+      vscode.TaskScope.Workspace,
+      taskName,
+      "firebase",
+      new vscode.ShellExecution(command, executionOptions),
     );
+    task.presentationOptions = presentationOptions;
+    await vscode.tasks.executeTask(task);
   });
 }
 
@@ -81,7 +88,9 @@ export function registerTerminalTasks(
     // TODO: optional debug mode
     runTerminalTask(
       "firebase emulators",
-      `${settings.firebasePath} emulators:start`,
+      `${settings.firebasePath} emulators:start --project ${currentProjectId.value}`,
+      // emulators:start almost never ask interactive questions.
+      { focus: false },
     );
   });